hw-6 debug bugs bugs bugs

+++ spring-filg/spring-hw-06/src/main/java/ru/otus/spring/repositories/BookRepository.java
@@ -18,7 +18,7 @@ public interface BookRepository {
 
     Book save(Book newBook);
 
-    int countBooks();
+    long countBooks();
 
     void update(Book book);
